Clean up tests (and find a bug in the JS implementation?)
